Rewrite Entities to MongoDB format. Rewire DAO to MongoDB host and rewrite its GetAll() Method

+++ src/dbobjects/entities/Entity.java
@@ -2,6 +2,8 @@ package dbobjects.entities;
 
 import dbobjects.DbObject;
 
+import java.math.BigInteger;
+
 public interface Entity extends DbObject {
-    int getId();
+    BigInteger getId();
 }
